introduce WasmModuleState
